+++ spec/integration/resource/api500/c7000/volume_attachment/create_spec.rb
@@ -1,0 +1,7 @@
+require 'spec_helper'
+
+klass = OneviewSDK::API500::C7000::VolumeAttachment
+RSpec.describe klass, integration: true, type: CREATE, sequence: seq(klass) do
+  let(:current_client) { $client_500 }
+  include_examples 'VolumeAttachmentCreateExample', 'integration api500 context'
+end