+++ spec/integration/shared_examples/volume_attachment/delete.rb
@@ -6,17 +6,16 @@
 #
 # Unless required by applicable law or agreed to in writing, software distributed
 # under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
-# CONDITIONS OF ANY KIND, either express or implied. See the License for the specific
-# language governing permissions and limitations under the License.
+# CONDITIONS OF ANY KIND, either express or implied. See the License for the
+# specific language governing permissions and limitations under the License.
 
-require_relative '../c7000/volume_attachment'
+RSpec.shared_examples 'VolumeAttachmentDeleteExample' do |context_name|
+  include_context context_name
 
-module OneviewSDK
-  module API500
-    module Synergy
-      # Storage Volume Attachment resource implementation for API500 Synergy
-      class VolumeAttachment < OneviewSDK::API500::C7000::VolumeAttachment
-      end
+  describe '#delete' do
+    it 'should throw unavailable exception' do
+      item = described_class.new(current_client)
+      expect { item.delete }.to raise_error(OneviewSDK::MethodUnavailable)
     end
   end
 end